[android] - AnnotationManager and InfoWindowManager refactor
